+++ JoinUs/src/main/java/com/northgatecode/joinus/mongodb/Gender.java
@@ -12,22 +12,22 @@ import java.util.List;
 @Entity(noClassnameStored = true)
 public class Gender {
     @Id
-    private long id;
+    private int id;
     private String name;
 
     public Gender() {
     }
 
-    public Gender(long id, String name) {
+    public Gender(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
